# swagger/modelgate/modelgate.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "modelgate/modelgate.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "ModelService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/massage/list": {
      "post": {
        "operationId": "ModelService_MessageList",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/modelgateMessageListResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "chat",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/modelgateChat"
            }
          }
        ],
        "tags": [
          "ModelService"
        ]
      }
    },
    "/model/chat": {
      "post": {
        "operationId": "ModelService_Chat",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/modelgateChatResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "chatBody",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/modelgateChatBody"
            }
          }
        ],
        "tags": [
          "ModelService"
        ]
      }
    },
    "/start/start": {
      "post": {
        "operationId": "ModelService_StartStart",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/modelgateStartResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "ModelService"
        ]
      }
    }
  },
  "definitions": {
    "modelgateChat": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "modelgateChatAnswer": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        }
      }
    },
    "modelgateChatBody": {
      "type": "object",
      "properties": {
        "chat": {
          "$ref": "#/definitions/modelgateChat"
        },
        "question": {
          "$ref": "#/definitions/modelgateChatQuestion"
        }
      }
    },
    "modelgateChatQuestion": {
      "type": "object",
      "properties": {
        "q": {
          "type": "string"
        }
      }
    },
    "modelgateChatResponse": {
      "type": "object",
      "properties": {
        "answer": {
          "$ref": "#/definitions/modelgateChatAnswer"
        }
      }
    },
    "modelgateMessage": {
      "type": "object",
      "properties": {
        "text": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "direction": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "modelgateMessageListResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/modelgateMessage"
          }
        }
      }
    },
    "modelgateStartResponse": {
      "type": "object",
      "properties": {
        "chat": {
          "$ref": "#/definitions/modelgateChat"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}